requirements.txt
numpy==1.24.2
openai==0.27.6
openai_whisper==20231117
PyAudioWPatch==0.2.12.6
webrtcvad==2.0.10
